Use new rule field disabled="1" for new fields, instead of commenting them out.
Add flag "new" to newly added messages, do not remove it on subsequent runs.

When making rule fuzzy, do not change it's hint, but state the new terms
following the fuzzy flag; this prevent it from being unfuzzied on next
run as before, and allows proper auto-unfuzzing in case the old terms
come back.