Added the code to make the missing directories.
The update-po.sh script must not fail if the storage dirs for the 'mo'
files do not exist.  As the mo dirs are the part of .gitignore, they
will not be recreated at target site unless the update-po.sh script
takes care that they are made.

This patch makes the missing dirs at the update-po.sh script runtime.